# docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
...
  model.ModerationAction:
    type: object
...
  model.PaginatedResponse:
    properties:
      data: {}
      has_more:
        type: boolean
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Threadlight social platform API
  title: Polaris API
  version: "0.1"
paths:
...
  /moderation/actions:
    get:
      description: Get paginated list of moderation actions
      parameters:
      - description: Number of results per page (default 20)
        in: query
        name: limit
        type: integer
      - description: Number of results to skip (default 0)
        in: query
        name: offset
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.PaginatedResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: List moderation actions
      tags:
      - moderation
    post:
      description: Create a new moderation action against a post
      parameters:
      - description: Moderation action details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.ModerationAction'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.ModerationAction'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Create moderation action
      tags:
      - moderation
  /moderation/actions/{id}:
    get:
      description: Get a moderation action by ID
      parameters:
      - description: Action ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ModerationAction'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Get moderation action
      tags:
      - moderation
  /moderation/actions/{id}/jurors:
    post:
      description: Add the authenticated user as a juror for a moderation action
      parameters:
      - description: Action ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Add juror
      tags:
      - moderation
  /moderation/actions/{id}/resolve:
    post:
      description: Resolve/close a jury for a moderation action
      parameters:
      - description: Action ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Resolve jury
      tags:
      - moderation
  /moderation/jurors/{panel_id}/vote:
    post:
      description: Submit a vote on a jury panel for a moderation action
      parameters:
      - description: Panel ID
        in: path
        name: panel_id
        required: true
        type: integer
      - description: Vote details
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Vote on jury panel
      tags:
      - moderation
...
swagger: "2.0"